Adds coordinator.organizations module
